Reducing the API surface of test-common and providing basic use documentation and advice.
