+++ e2e/tests/test.rs
@@ -4,30 +4,31 @@ use test_common::*;
 #[test]
 fn create_drop_promscale_extension() {
     let _ = pretty_env_logger::try_init();
+    let pg_blueprint = PostgresContainerBlueprint::new();
+    let test_pg_instance = new_test_container_instance(&pg_blueprint);
+    let mut test_conn = test_pg_instance.connect();
 
-    PostgresTestConnection::in_docker_db(|test_conn| {
-        let result = test_conn
-            .client
-            .simple_query("CREATE EXTENSION promscale;")
-            .unwrap();
-        assert_eq!(result.len(), 1);
+    let result = test_conn
+        .simple_query("CREATE EXTENSION promscale;")
+        .unwrap();
+    assert_eq!(result.len(), 1);
 
-        let result = test_conn
-            .client
-            .simple_query("DROP EXTENSION promscale CASCADE;")
-            .unwrap();
+    let result = test_conn
+        .simple_query("DROP EXTENSION promscale CASCADE;")
+        .unwrap();
 
-        assert_eq!(result.len(), 1);
-    });
+    assert_eq!(result.len(), 1);
 }
 
 #[test]
 fn upgrade_promscale_extension_all_versions() {
     let _ = pretty_env_logger::try_init();
-    PostgresTestConnection::in_docker_db(|test_conn| {
-        // This query gets all possible upgrade paths for the extension. An upgrade
-        // path looks like: "0.1.0--0.2.0--0.2.1--0.3.0".
-        let path_rows = test_conn.client
+    let pg_blueprint = PostgresContainerBlueprint::new();
+    let test_pg_instance = new_test_container_instance(&pg_blueprint);
+    let mut test_conn = test_pg_instance.connect();
+    // This query gets all possible upgrade paths for the extension. An upgrade
+    // path looks like: "0.1.0--0.2.0--0.2.1--0.3.0".
+    let path_rows = test_conn
         .query(r#"
             SELECT path
             FROM pg_extension_update_paths('promscale')
@@ -42,56 +43,53 @@ fn upgrade_promscale_extension_all_versions() {
             "#, &[])
         .unwrap();
 
-        let version_paths: Vec<Vec<String>> = path_rows
-            .iter()
-            .map(|r| {
-                let path = r.get::<&str, &str>("path");
-                // Split string "0.1.0--0.2.0" into vec ["0.1.0", "0.2.0"].
-                path.split("--")
-                    .map(str::to_string)
-                    .collect::<Vec<String>>()
-            })
-            .collect();
+    let version_paths: Vec<Vec<String>> = path_rows
+        .iter()
+        .map(|r| {
+            let path = r.get::<&str, &str>("path");
+            // Split string "0.1.0--0.2.0" into vec ["0.1.0", "0.2.0"].
+            path.split("--")
+                .map(str::to_string)
+                .collect::<Vec<String>>()
+        })
+        .collect();
 
-        for version_path in version_paths {
-            let mut prev_version: Option<String> = None;
-            for version in version_path {
-                match prev_version {
-                    None => {
-                        info!("Creating extension at version {}", version);
-                        let res = test_conn.client.query(
-                            &format!("CREATE EXTENSION promscale VERSION '{}'", version),
-                            &[],
-                        );
-                        assert!(
-                            res.is_ok(),
-                            "cannot create extension at version {}",
-                            version
-                        );
-                    }
-                    Some(prev_version) => {
-                        info!(
-                            "Upgrading extension from version {} to {}",
-                            prev_version, version
-                        );
-                        let res = test_conn.client.query(
-                            &format!("ALTER EXTENSION promscale UPDATE TO '{}'", version),
-                            &[],
-                        );
-                        assert!(
-                            res.is_ok(),
-                            "cannot upgrade extension from version {} to {}",
-                            prev_version,
-                            version
-                        );
-                    }
+    for version_path in version_paths {
+        let mut prev_version: Option<String> = None;
+        for version in version_path {
+            match prev_version {
+                None => {
+                    info!("Creating extension at version {}", version);
+                    let res = test_conn.query(
+                        &format!("CREATE EXTENSION promscale VERSION '{}'", version),
+                        &[],
+                    );
+                    assert!(
+                        res.is_ok(),
+                        "cannot create extension at version {}",
+                        version
+                    );
                 }
-                prev_version = Some(version);
+                Some(prev_version) => {
+                    info!(
+                        "Upgrading extension from version {} to {}",
+                        prev_version, version
+                    );
+                    let res = test_conn.query(
+                        &format!("ALTER EXTENSION promscale UPDATE TO '{}'", version),
+                        &[],
+                    );
+                    assert!(
+                        res.is_ok(),
+                        "cannot upgrade extension from version {} to {}",
+                        prev_version,
+                        version
+                    );
+                }
             }
-            let res = test_conn
-                .client
-                .simple_query("DROP EXTENSION promscale CASCADE;");
-            assert!(res.is_ok(), "cannot drop extension");
+            prev_version = Some(version);
         }
-    });
+        let res = test_conn.simple_query("DROP EXTENSION promscale CASCADE;");
+        assert!(res.is_ok(), "cannot drop extension");
+    }
 }
